simpler solution, just need to generate the random str to add to the css and js now
