+++ bin/util
@@ -6,7 +6,6 @@ const fs = require('fs');
 const path = require('path');
 const moment = require('moment');
 const es = require('../db/elasticsearch');
-const sa = require('../db/cache/static-assets');
 
 const searchConfig = require('../config/search.js');
 const formatHelper = require('../helpers/format');
@@ -136,7 +135,7 @@ async function indexVillagers(es) {
             species: parsed.species,
             name: parsed.name,
             url: urlHelper.getEntityUrl(searchConfig.villagerEntityType, parsed.id),
-            imageUrl: sa.computeStaticAssetUrl(urlHelper.getImageUrl(searchConfig.villagerEntityType,
+            imageUrl: urlHelper.computeStaticAssetUrl(urlHelper.getImageUrl(searchConfig.villagerEntityType,
                 urlHelper.THUMB, parsed.id)),
             game: [],
             personality: [],
@@ -191,7 +190,7 @@ async function indexItems(es) {
             category: parsed.category,
             game: [],
             url: urlHelper.getEntityUrl(searchConfig.itemEntityType, parsed.id),
-            imageUrl: sa.computeStaticAssetUrl(urlHelper.getImageUrl(searchConfig.itemEntityType, urlHelper.THUMB,
+            imageUrl: urlHelper.computeStaticAssetUrl(urlHelper.getImageUrl(searchConfig.itemEntityType, urlHelper.THUMB,
                 parsed.id))
         };
 
@@ -261,17 +260,6 @@ if (!command) {
         })
         .catch((e) => {
             console.log('Failed to build search index.');
-            console.error(e);
-            process.exit(1);
-        });
-} else if (command === 'delete-asset-cache') {
-    sa.flushCache()
-        .then(() => {
-            console.log('Static asset cache flushed successfully.');
-            process.exit(0);
-        })
-        .catch((e) => {
-            console.log('Failed to flush static asset cache.');
             console.error(e);
             process.exit(1);
         });
