+++ helpers/middleware/app-state.js
@@ -1,6 +1,13 @@
-const sa = require('../../db/cache/static-assets')
+const urlHelper = require('../../helpers/url')
 
-const doWork = async (req, res) => {
+/**
+ * Basic middleware that populates some user state data so that every template can access it, if needed.
+ *
+ * @param req
+ * @param res
+ * @param next
+ */
+module.exports = (req, res, next) => {
     // Google Analytics ID
     if (process.env.GOOGLE_ANALYTICS_ID) {
         res.locals.gaId = process.env.GOOGLE_ANALYTICS_ID;
@@ -16,19 +23,7 @@ const doWork = async (req, res) => {
     }
 
     // Stylesheet and JavaScript URL.
-    res.locals.stylesheetUrl = await sa.getStaticAssetUrl('/stylesheets/style.css');
-    res.locals.javascriptUrl = await sa.getStaticAssetUrl('/javascripts/bundle.js');
-};
-
-/**
- * Basic middleware that populates some user state data so that every template can access it, if needed.
- *
- * @param req
- * @param res
- * @param next
- */
-module.exports = (req, res, next) => {
-    doWork(req, res)
-        .then(next)
-        .catch(next);
+    res.locals.stylesheetUrl = urlHelper.getCacheBustedUrl('/stylesheets/style.css');
+    res.locals.javascriptUrl = urlHelper.getCacheBustedUrl('/javascripts/bundle.js');
+    next();
 };
